Add files via upload
The project was originally a class assignment, with each iteration done by the end of every other week on top of learning new concepts to apply to the final website. I intend on making changes based on the JavaScript classes I am taking, and by learning more about web design. Started on the project from Feb-April.